search condor_history backwards (default)
Searching -forwards can be very slow for keepup, presumably because
the schedd has to search through the entire history from the
beginning to find jobs matching the constraint.

There are two downsides:
1. Can't populate back history, beyond <limit> jobs on the first run.
2. We'll only get the latest <limit> jobs each run, and the checkpoint
   will get updated to the last. So, if more than <limit> jobs
   finish between runs, we'll miss them.

We should use the -since and/or -completedsince options in 8.5/8.6,
although many production pools are still on 8.4. Maybe we can require
8.6 on the client and it will degrade gracefully talking to an older
schedd?